% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closedpop_parms.R
\name{closedpop_parms}
\alias{closedpop_parms}
\title{A closed population parameters}
\usage{
closedpop_parms(maxage, Lmat, Lfish, M, Fi, Linf, k, a0, pW, qW)
}
\arguments{
\item{maxage:}{max age of the species ie. number of age classes}

\item{Lmat:}{length at maturity}

\item{M:}{the natural mortality rate, if unknown generally use 0.2}

\item{Fi:}{the fishing mortality rate, F, find in stock assessment if don't have more localized estimate}

\item{Linf:}{von-bertallanfy growth parameter estimate, can find on fishbase}

\item{k:}{von-bertallanfy growth parameter estimate}

\item{a0:}{same as t0 in von-bertallanfy growth parameter}

\item{pw:}{weight length relationship estimate, same as a on fishbase.org but need to divide by 1000 to get in kg not grams}

\item{qw:}{weight length relationship estmate, same as b on fishbase.org}
}
\description{
This function calculates the ratio change in a fished poulation after a marine protected area is implemented assuming  closed population.
It also calculates parameters of the transient response of an MPA using the methods described in
White et al. 2013 'Transient responses of fished populations to marine reserve establishment' published in conservation letters.
It uses a Leslies matrix and the output is a data frame with population ratio changes from a fished population to unfished in the MPA.
Lambda is set to 1 inside the MPA, the starting population vector is the stable age distribution of the fished state
}
\examples{
closedpop_parms(maxage=25,Lmat=18,Lfish=25,M=0.2,Fi=0.17, Linf=37.8,k=0.23,a0=-0.7,pW=6.29e-06,qW=3.172)
closedpop_parms()
}
\keyword{Leslie}
\keyword{closed}
\keyword{dynamics,}
\keyword{matrix}
\keyword{population}
\keyword{population,}
