% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closedpop_time.R
\name{closedpop_time}
\alias{closedpop_time}
\title{A closed population function that calculate the lenght of time of the transient period after MPA implementation for a deterministic model
This function calculates the ratio change in a fished poulation after a marine protected area is implemented assuming  closed population
method is from White et al. 2013 'transient responses of fished populations to marine reserve establishment published in conservation letters
assumes a lambda of 1.0 in the MPA}
\usage{
closedpop_time(maxage, Lmat, Lfish, M, Fi, Linf, k, a0, pW, qW)
}
\arguments{
\item{maxage}{max age of the species ie. number of age classes}

\item{Lmat}{length at maturity}

\item{M}{the natural mortality rate, if unknown generally use 0.2}

\item{Fi}{the fishing mortality rate, F, find in stock assessment if don't have more localized estimate}

\item{Linf}{von-bertallanfy growth parameter estimate, can find on fishbase.org}

\item{k}{von-bertallanfy growth parameter estimate}

\item{a0}{same as t0 in von-bertallanfy growth parameter}

\item{pw}{weight length relationship estimate, same as a on fishbase.org but need to divide by 1000 to get in kg not grams}

\item{qw}{weight length relationship estmate, same as b on fishbase.org}
}
\description{
A closed population function that calculate the lenght of time of the transient period after MPA implementation for a deterministic model
This function calculates the ratio change in a fished poulation after a marine protected area is implemented assuming  closed population
method is from White et al. 2013 'transient responses of fished populations to marine reserve establishment published in conservation letters
assumes a lambda of 1.0 in the MPA
}
\examples{
closedpop_time(M=0.2,Fi=0.14,Lfish=25,Linf=37.8,k=0.13,a0=-0.7,maxage=25,pW=9.37e-06,qW=3.172)
closed_time()
}
\keyword{Leslie}
\keyword{closed}
\keyword{dynamics,}
\keyword{matrix}
\keyword{population}
\keyword{population,}
