% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openpop_plots.R
\name{openpop_plot}
\alias{openpop_plot}
\title{An open population calculation function}
\usage{
openpop_plot(tf, maxage, M, Fi, Lfish, Linf, k, a0, pW, qW, R, sig_r,
  simulations, MPAtime)
}
\arguments{
\item{tf:}{the time steps to run the population}

\item{maxage:}{max age of the species ie. number of age classes}

\item{M:}{the natural mortality rate, if unknown generally use 0.2}

\item{Fi:}{the fishing mortality rate, F, find in stock assessment if don't have more localized estimate}

\item{Lfish:}{length at first capture in fishery}

\item{Linf:}{asymptotic growth rate used in von-Bertallanfy growth equation}

\item{k:}{von-bertallanfy growth parameter estimate}

\item{a0:}{age at length zero used in von-Bertallanfy growth equation}

\item{pW:}{weight length relationship parameter, same as a on fishbase.org but need to divide by 1000 to get in kg not grams}

\item{qW:}{weight length relationship parameter, same as b on fishbase.org}

\item{R:}{number of recruits to enter the population}

\item{sig_r:}{stochastic parameter, log-normal distribution, around recruitment}

\item{simulations:}{the number of simulations to run the population with stochastic recruitment}

\item{MPAtime:}{the timestep that the MPA is implemented}
}
\description{
This function calculates the ratio change abundance and biomass of a fished population, after a marine protected area is implemented assuming a population with external recruitment.
The output is three plots, one with the deterministic ratio changes in abundance and biomass for fished age classes.
The second is the abundance ratio change with stochastic recruitment.
The third is the biomass ratio change with stochastic recruitment.
}
\examples{
openpop_plot(tf=50,M=0.2,Fi=0.14,Lfish=25,Linf=37.8,k=0.13,a0=-0.7,
maxage=25,pW=9.37e-06,qW=3.172,sig_r=0.3, R=500, MPAtime=1,simulations=100)
}
\keyword{dynamics}
\keyword{open}
\keyword{population}
\keyword{population,}
