% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openpop_plots.R
\name{openpop_plots}
\alias{openpop_plots}
\title{An open population plot function}
\usage{
openpop_plots(tf, maxage, M, Fi, Lfish, Linf, k, a0, pW, qW, R, sig_r,
  iterations, MPAtime)
}
\arguments{
\item{tf:}{the total number of time steps to run the model}

\item{maxage:}{max age of the species ie. number of age classes}

\item{M:}{the natural mortality rate, if unknown generally use 0.2}

\item{Fi:}{the fishing mortality rate, F, find in stock assessment if don't have more localized estimate}

\item{Linf:}{von-bertallanfy growth parameter estimate, can find on fishbase}

\item{k:}{von-bertallanfy growth parameter estimate}

\item{a0:}{same as t0 in von-bertallanfy growth parameter}

\item{pw:}{weight length relationship estimate, same as a on fishbase.org but need to divide by 1000 to get in kg not grams}

\item{qw:}{weight length relationship estmate, same as b on fishbase.org}

\item{R:}{number of recruits entering the population}

\item{sig_r:}{stochastic parameter, log-normal distribution, around recruitment}

\item{iterations:}{number of stochastic simulations to run to calculated 95% confidence intervals}

\item{MPAtime:}{the time step when the MPA is implemented in the model}
}
\description{
This function calculates the ratio change abundance and biomass of a fished poulation, after a marine protected area is implemented assuming a population with external recruitment.
The output is 4 plots with population ratio changes from a fished population to unfished in the MPA
These include abundance and biomass ratio changes with the deterministic model and with stochastic recruitment
It includes output using stochastic recruitment
the function requires the use of the packages ggplot2 and cowplot, these must be installed to use this function
}
\examples{
#openpop_plots(tf=50,M=0.2,Fi=0.14,Lfish=25,Linf=37.8,k=0.13,a0=-0.7,maxage=25,pW=9.37e-06,qW=3.172,R=500,sig_r=0.5, iterations=10,MPAtime=5)
openpop_plots()
}
\keyword{open}
\keyword{population}
