% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openpop_ratio.R
\name{openpop_ratio}
\alias{openpop_ratio}
\title{An open population ratio calculation function}
\usage{
openpop_ratio(maxage, M, Fi, Lfish, Linf, k, a0, pW, qW, sig_r)
}
\arguments{
\item{maxage:}{max age of the species ie. number of age classes}

\item{Lmat:}{length at maturity}

\item{M:}{the natural mortality rate, if unknown generally use 0.2}

\item{Fi:}{the fishing mortality rate, F, find in stock assessment if don't have more localized estimate}

\item{Linf:}{von-bertallanfy growth parameter estimate, can find on fishbase}

\item{k:}{von-bertallanfy growth parameter estimate}

\item{a0:}{same as t0 in von-bertallanfy growth parameter}

\item{pw:}{weight length relationship estimate, same as a on fishbase.org but need to divide by 1000 to get in kg not grams}

\item{qw:}{weight length relationship estmate, same as b on fishbase.org}

\item{sig_r:}{stochastic parameter, log-normal distribution, around recruitment}
}
\description{
This function calculates the ratio change abundance and biomass of a fished poulation, after a marine protected area is implemented assuming a population with external recruitment.
The output is a data frame with population ratio changes from a fished population to unfished in the MPA
It includes output using stochastic recruitment
}
\examples{
openpop_ratio(M=0.2,Fi=0.14,Lfish=25,Linf=37.8,k=0.13,a0=-0.7,maxage=25,pW=9.37e-06,qW=3.172,sig_r=0.5)
openpop_ratio()
}
\keyword{Leslie}
\keyword{closed}
\keyword{dynamics,}
\keyword{matrix}
\keyword{population}
\keyword{population,}
