% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openpop_ROC.R
\name{openpop_ROC}
\alias{openpop_ROC}
\title{An open population function to calculate population changes in a marine protected area, MPA}
\usage{
openpop_ROC(tf, maxage, M, Fi, Lfish, Linf, k, a0, pW, qW, R, sig_r, MPAtime,
  simulations)
}
\arguments{
\item{tf:}{the time steps to run the population}

\item{maxage:}{max age of the species ie. number of age classes}

\item{Lmat:}{length at maturity}

\item{M:}{the natural mortality rate}

\item{Fi:}{the fishing mortality rate, F, find in stock assessment if don't have more localized estimate}

\item{Linf:}{asymptotic growth rate used in von-Bertallanfy growth equation, can find on fishbase}

\item{k:}{von-bertallanfy growth parameter estimate}

\item{a0:}{the age at length 0 used in the von-Bertallanfy growth equation}

\item{pW:}{weight length relationship parameter, same as a on fishbase.org but need to divide by 1000 to get in kg not grams}

\item{qW:}{weight length relationship parameter, same as b on fishbase.org}

\item{R:}{number of recruits entering the population}

\item{sig_r:}{stochastic parameter, log-normal distribution, around recruitment}

\item{MPAtime:}{the time step to implement the MPA}

\item{simulations:}{the number of simulations to run}
}
\value{
roc2.a: returns the area under an reciever operating characteristic curve for abundance at year 2 since MPA implementation

roc5.a: returns the area under an reciever operating characteristic curve for abundance at year 5 since MPA implementation

roc10.a: returns the area under an reciever operating characteristic curve for abundance at year 10 since MPA implementation

roc20.a: returns the area under an reciever operating characteristic curve for abundance at year 20 since MPA implementation

roc2.b: returns the area under an reciever operating characteristic curve for biomass at year 2 since MPA implementation

roc5.b: returns the area under an reciever operating characteristic curve for biomass at year 5 since MPA implementation

roc10.b: returns the area under an reciever operating characteristic curve for biomass at year 10 since MPA implementation

roc20.b: returns the area under an reciever operating characteristic curve for biomass at year 20 since MPA implementation
}
\description{
This function calculates the ratio change abundance and biomass of a fished poulation, after a marine protected area is implemented assuming a population with external recruitment.
The output is a data frame with population ratio changes from a fished population to unfished in the MPA
It includes deterministic population model output and output using stochastic recruitment
}
\examples{
openpop_ratio(tf=50, M=0.2,Fi=0.14,Lfish=25,Linf=37.8,k=0.13,a0=-0.7,maxage=25,pW=9.37e-06,qW=3.172,R=500,
 sig_r=0.5, MPAtime=1,simulations=100)
}
\keyword{dynamics}
\keyword{open}
\keyword{population}
\keyword{population,}
