% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openpop_time.R
\name{openpop_time}
\alias{openpop_time}
\title{An open population function that calculate the length of time to reach equilibrium in response to an MPA}
\usage{
openpop_time(maxage, M, Fi, Lfish, Linf, k, a0, pW, qW)
}
\arguments{
\item{maxage:}{max age of the species ie. number of age classes}

\item{Lmat:}{length at maturity}

\item{M:}{the natural mortality rate, if unknown generally use 0.2}

\item{Fi:}{the fishing mortality rate, F, find in stock assessment if don't have more localized estimate}

\item{Linf:}{von-bertallanfy growth parameter estimate, can find on fishbase}

\item{k:}{von-bertallanfy growth parameter estimate}

\item{a0:}{same as t0 in von-bertallanfy growth parameter}

\item{pw:}{weight length relationship estimate, same as a on fishbase.org but need to divide by 1000 to get in kg not grams}

\item{qw:}{weight length relationship estmate, same as b on fishbase.org}

\item{sig_r:}{stochastic parameter, log-normal distribution, around recruitment}
}
\description{
This function calculates the ratio change in a fished poulation after a marine protected area is implemented assuming external recruitment.
The output is the time to reach the abundance and biomass equilibrium.
assumes a lambda of 1.0 in the MPA and calculate change from the stable age distribution of the fished population
}
\examples{
openpop_time(M=0.2,Fi=0.14,Lfish=25,Linf=37.8,k=0.13,a0=-0.7,maxage=25,pW=9.37e-06,qW=3.172)
openpop_time()
}
\keyword{Leslie}
\keyword{closed}
\keyword{dynamics,}
\keyword{matrix}
\keyword{population}
\keyword{population,}
